requirements.txt
pyspark
kafka-python
pymongo
psycopg2-binary
sqlalchemy
python-dotenv
Flask==2.1.2
